+++ PSBlueSky.psm1
@@ -38,6 +38,8 @@ $OnRemoveScript = {
         $script:PSCmd.Runspace.Close()
         $script:PSCmd.Runspace.Dispose()
     }
+    #clean up type data to avoid errors on re-importing
+    "PSBlueskySession","PSBlueskySearchResult","PSBlueskyProfile","PSBlueskyFollowProfile" | Remove-TypeData
 }
 
 $ExecutionContext.SessionState.Module.OnRemove += $OnRemoveScript
@@ -51,10 +53,10 @@ Update-TypeData -TypeName 'PSBlueskySession' -MemberType AliasProperty -MemberNa
 Update-TypeData -TypeName 'PSBlueskySession' -MemberType AliasProperty -MemberName RefreshToken -Value RefreshJwt -Force
 Update-TypeData -TypeName 'PSBlueskySession' -MemberType ScriptProperty -MemberName Age -Value { (Get-Date) - $this.Date } -Force
 Update-TypeData -TypeName 'PSBlueskySession' -MemberType ScriptMethod -MemberName Refresh -Value { Update-BskySession -RefreshToken $this.RefreshJwt } -Force
-
 Update-TypeData -TypeName 'PSBlueskySearchResult' -MemberType ScriptProperty -MemberName Age -Value { (Get-Date) - $this.Created } -Force
 Update-TypeData -TypeName 'PSBlueskyProfile' -MemberType ScriptProperty -MemberName Age -Value { (Get-Date) - $this.Created } -Force
 Update-TypeData -TypeName 'PSBlueskyFollowProfile' -MemberType ScriptProperty -MemberName Age -Value { (Get-Date) - $this.Created } -Force
+
 #endregion
 
 #region verbose command highlighting